fix anonotataions

+++ 0x01-python_async_function/2-measure_runtime.py
@@ -7,7 +7,7 @@ wait_n = __import__('1-concurrent_coroutines').wait_n
 import asyncio
 
 
-def measure_time(n: int, max_delay: float) -> float:
+def measure_time(n: int, max_delay: int) -> float:
     """
     measures the runtime of wait_n function
     """
